Extract common repository methods in base class
