+++ sqlalchemy_bind_manager/_repository/sync.py
@@ -1,24 +1,20 @@
 from abc import ABC
-from typing import Union, Generic, Type, Iterable, Tuple
+from typing import Union, Generic, Iterable, Tuple
 
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import Session, object_mapper, Mapper, class_mapper
-from sqlalchemy.orm.exc import UnmappedInstanceError
-from sqlalchemy.sql import Select
+from sqlalchemy.orm import Session
 
 from .._bind_manager import SQLAlchemyBindManager, DEFAULT_BIND_NAME, SQLAlchemyBind
 from ..exceptions import (
     ModelNotFound,
-    UnmappedProperty,
-    InvalidModel, UnsupportedBind,
+    UnsupportedBind,
 )
-from .common import MODEL, PRIMARY_KEY, SortDirection
+from .common import MODEL, PRIMARY_KEY, SortDirection, BaseRepository
 
 
-class SQLAlchemySyncRepository(Generic[MODEL], ABC):
+class SQLAlchemySyncRepository(Generic[MODEL], BaseRepository[MODEL], ABC):
     _session: Session
-    _model: Type[MODEL]
 
     def __init__(
         self, sa_manager: SQLAlchemyBindManager, bind_name: str = DEFAULT_BIND_NAME
@@ -124,77 +120,3 @@ class SQLAlchemySyncRepository(Generic[MODEL], ABC):
         except:
             session.rollback()
             raise
-
-    def _is_mapped_object(self, obj: object) -> bool:
-        """Checks if the object is handled by the repository and is mapped in SQLAlchemy.
-
-        :param obj: a mapped object instance
-        :return: True if the object is mapped and matches self._model type, False if it's not a mapped object
-        :rtype: bool
-        :raises InvalidModel: when the object is mapped but doesn't match self._model type
-        """
-        # TODO: This is probably redundant, we could do these checks once in __init__
-        try:
-            object_mapper(obj)
-            if isinstance(obj, self._model):
-                return True
-            raise InvalidModel(
-                f"This repository can handle only `{self._model}` models. `{type(obj)}` has been passed."
-            )
-        except UnmappedInstanceError:
-            return False
-
-    def _validate_mapped_property(self, property_name: str) -> None:
-        """Checks if a property is mapped in the model class.
-
-        :param property_name: The name of the property to be evaluated.
-        :type property_name: str
-        :raises UnmappedProperty: When the property is not mapped.
-        """
-        m: Mapper = class_mapper(self._model)
-        if property_name not in m.column_attrs:  # type: ignore
-            raise UnmappedProperty(
-                f"Property `{property_name}` is not mapped in the ORM for model `{self._model}`"
-            )
-
-    def _filter_select(self, stmt: Select, **search_params) -> Select:
-        """Build the query filtering clauses from submitted parameters.
-
-        E.g.
-        _filter_select(stmt, name="John") adds a `WHERE name = John` statement
-
-        :param stmt: a Select statement
-        :type stmt: Select
-        :param search_params: Any keyword argument to be used as equality filter
-        :return: The filtered query
-        """
-        # TODO: Add support for offset/limit
-        # TODO: Add support for relationship eager load
-        for k, v in search_params.items():
-            self._validate_mapped_property(k)
-            stmt = stmt.where(getattr(self._model, k) == v)
-        return stmt
-
-    def _filter_order_by(
-        self, stmt: Select, order_by: Iterable[Union[str, Tuple[str, SortDirection]]]
-    ) -> Select:
-        """Build the query ordering clauses from submitted parameters.
-
-        E.g.
-        _filter_order_by(stmt, ['name']) adds a `ORDER BY name` statement
-        _filter_order_by(stmt, [('name', SortDirection.ASC)]) adds a `ORDER BY name ASC` statement
-
-        :param stmt: a Select statement
-        :type stmt: Select
-        :param order_by: a list of columns, or tuples (column, direction)
-        :return: The filtered query
-        """
-        for value in order_by:
-            if isinstance(value, str):
-                self._validate_mapped_property(value)
-                stmt = stmt.order_by(getattr(self._model, value))
-            else:
-                self._validate_mapped_property(value[0])
-                stmt = stmt.order_by(value[1].value(getattr(self._model, value[0])))
-
-        return stmt
